MRG: add skipmer capacity to sourmash python layer via ffi (#3446)
This PR updates the FFI and python layer to allow the skipmer moltypes
(`skipm1n3`, `skipm2n3`). We will keep this as an undocumented
experimental feature for now. There are no guarantees at the moment that
skipmers will work with all sourmash commands, as there are no explicit
tests in place. There *are* tests for using skipmers with branchwater,
so all skipmer searching is best done in the plugin for now. This PR
enables critical handy utilities, though: `sig cat`, `sig summarize`,
`sig describe`, etc.

Documentation and additional tests should be added prior to
release(#3449).

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: C. Titus Brown <titus@idyll.org>